requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.0
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
cachetools==5.2.0
cairocffi==1.4.0
captum==0.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.6
coral-pytorch==1.4.0
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
et-xmlfile==1.1.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.1
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.4.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
optuna==3.0.4
packaging==22.0
pandas==1.5.2
pbr==5.11.0
Pillow==9.3.0
prettytable==3.5.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytorch-lightning==1.8.5.post0
pytz==2022.6
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.45
stevedore==4.1.1
tbparse==0.0.7
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tinycss2==1.2.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchinfo==1.7.1
torchmetrics==0.11.0
torchvision==0.14.1+cu116
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
